requirements.txt
langchain
langchain-core 
langchain-openai 
langchain-community 
openai 
langchain-chroma 
langchain-experimental 
chromadb 
tiktoken
python-dotenv
streamlit
fastapi
uvicorn[standard]
gunicorn
chromadb
openai
loguru
psycopg2-binary